ci: add validation workflow (#3)
this adds a github workflow for validation, which runs format, lint,
test, and e2e tests against a production build.

currently this runs for all packages in the monorepo, which is fine for
now, but we may need to reconsider once we manage the dariah.eu website
package here as well. this will also require rethinking our naming
conventions for environment variables for different app packages.

runs on pull requests and pushes to `main`.