feat: add default layout for landing page (#12)
this adds a basic default layout for landing page, imprint page, and
contact page. no mobile design yet. this is mainly to make the analytics
e2e test work again, which assumes an imprint link on the page.

also adds placeholder pages for privacy-policy and terms-of-use pages,
and an initial dashboard layout using a intentui sidebar template.

entry point for dashboard pages:
http://localhost:3000/en/dashboard/administrator

not related: ensures that dashboard pages are not listed in sitemap.