feat: add seed scripts to package exports (#55)
this adds all seed scripts to package exports, and returns a seed
manifest from the image-service package, which can be fed into the db
seed script, so we can populate the db with object keys which actually
exist in the s3 object store.

also changes the status of seeded entities to "published" instead of
"draft", so we can use them to test the api server (which only returns
published resources).