chore: create utility package to orchestrate seeding (#57)
this creates a new utility package (in the `lib` folder) to orchestrate
seeding scripts for dev services. in particular, we want to be able to
seed the s3 object store with images, and pass the keys to the database
seed script, so we can reference actually existing assets. script can be
invoked with `pnpm dev:seed` from the monorepo root.